[Metabot] Remove non-streaming metabot code (#61988)
* remove streaming code

* remove comment
