+++ enterprise/frontend/src/metabase-enterprise/metabot/metabot.unit.spec.tsx
@@ -16,7 +16,13 @@ import {
 import { logout } from "metabase/auth/actions";
 import * as domModule from "metabase/lib/dom";
 import { useRegisterMetabotContextProvider } from "metabase/metabot";
-import type { SuggestedMetabotPrompt, User } from "metabase-types/api";
+import {
+  type MockStreamedEndpointParams,
+  createMockReadableStream,
+  createPauses,
+  mockStreamedEndpoint,
+} from "metabase-enterprise/api/ai-streaming/test-utils";
+import type { User } from "metabase-types/api";
 import {
   createMockTokenFeatures,
   createMockUser,
@@ -34,18 +40,24 @@ import { useMetabotAgent } from "./hooks";
 import {
   type MetabotState,
   addUserMessage,
+  getHistory,
+  getMetabot,
   getMetabotConversationId,
   getMetabotInitialState,
+  getMetabotState,
   metabotReducer,
   setVisible,
 } from "./state";
+
+const mockAgentEndpoint = (params: MockStreamedEndpointParams) =>
+  mockStreamedEndpoint("/api/ee/metabot-v3/v2/agent-streaming", params);
 
 function setup(
   options: {
     ui?: React.ReactElement;
     metabotPluginInitialState?: MetabotState;
     currentUser?: User | null | undefined;
-    promptSuggestions?: SuggestedMetabotPrompt[];
+    promptSuggestions?: { prompt: string }[];
   } | void,
 ) {
   const settings = mockSettings({
@@ -62,7 +74,7 @@ function setup(
     metabotPluginInitialState = {
       ...getMetabotInitialState(),
       visible: true,
-      useStreaming: false,
+      useStreaming: true,
     },
     promptSuggestions = [],
   } = options || {};
@@ -95,8 +107,8 @@ const input = () => screen.findByTestId("metabot-chat-input");
 const enterChatMessage = async (message: string, send = true) =>
   userEvent.type(await input(), `${message}${send ? "{Enter}" : ""}`);
 const closeChatButton = () => screen.findByTestId("metabot-close-chat");
-const resetChatButton = () => screen.findByTestId("metabot-reset-chat");
 const responseLoader = () => screen.findByTestId("metabot-response-loader");
+const resetChatButton = () => screen.findByTestId("metabot-reset-chat");
 
 const assertVisible = async () =>
   expect(await screen.findByTestId("metabot-chat")).toBeInTheDocument();
@@ -109,15 +121,32 @@ const assertNotVisible = async () =>
 const hideMetabot = (dispatch: any) => act(() => dispatch(setVisible(false)));
 const showMetabot = (dispatch: any) => act(() => dispatch(setVisible(true)));
 
-const getMetabotState = (store: any) => store.getState().plugins.metabotPlugin;
-
-const lastReqBody = async () => {
-  const lastCall = fetchMock.lastCall();
-  const bodyStr = String(await lastCall?.[1]?.body) || "";
-  return JSON.parse(bodyStr);
+const assertConversation = async (
+  expectedMessages: ["user" | "agent", string][],
+) => {
+  if (!expectedMessages.length) {
+    await waitFor(() => {
+      expect(
+        screen.queryByTestId("metabot-chat-message"),
+      ).not.toBeInTheDocument();
+    });
+  } else {
+    const realMessages = await chatMessages();
+    expect(expectedMessages.length).toBe(realMessages.length);
+    expectedMessages.forEach(([expectedRole, expectedMessage], index) => {
+      const realMessage = realMessages[index];
+      expect(realMessage).toHaveAttribute("data-message-role", expectedRole);
+      expect(realMessage).toHaveTextContent(expectedMessage);
+    });
+  }
 };
 
-describe("metabot", () => {
+const lastReqBody = async (agentSpy: ReturnType<typeof mockAgentEndpoint>) => {
+  await waitFor(() => expect(agentSpy).toHaveBeenCalled());
+  return JSON.parse(agentSpy.mock.lastCall?.[1]?.body as string);
+};
+
+describe("metabot-streaming", () => {
   describe("ui", () => {
     it("should be able to render metabot", async () => {
       setup();
@@ -133,10 +162,7 @@ describe("metabot", () => {
 
     it("should show empty state ui if conversation is empty", async () => {
       setup();
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-      );
+      mockAgentEndpoint({ textChunks: whoIsYourFavoriteResponse });
 
       expect(
         await screen.findByTestId("metabot-empty-chat-info"),
@@ -219,30 +245,9 @@ describe("metabot", () => {
       }
     });
 
-    it("should render markdown for metabot's replies", async () => {
-      setup();
-      fetchMock.post(`path:/api/ee/metabot-v3/v2/agent`, {
-        ...whoIsYourFavoriteResponse,
-        reactions: [
-          {
-            type: "metabot.reaction/message",
-            message: "# You are... but don't tell anyone!",
-          },
-        ],
-      });
-      await enterChatMessage("Who is your favorite?");
-
-      const heading = await screen.findByRole("heading", { level: 1 });
-      expect(heading).toBeInTheDocument();
-      expect(heading).toHaveTextContent(`You are... but don't tell anyone!`);
-    });
-
     it("should not render markdown for user messages", async () => {
       setup();
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-      );
+      mockAgentEndpoint({ textChunks: whoIsYourFavoriteResponse });
 
       const msg = "# Who is your favorite?";
       await enterChatMessage(msg);
@@ -252,89 +257,80 @@ describe("metabot", () => {
       ).not.toBeInTheDocument();
     });
 
-    it("should provide prompt suggestions if available", async () => {
-      const prompts = [
-        {
-          id: 1,
-          metabot_id: 1,
-          prompt: "What is the total revenue for this quarter?",
-          model: "metric" as const,
-          model_id: 1,
-          model_name: "Quarterly Revenue Calculator",
-          created_at: "2025-05-15T10:30:00Z",
-          updated_at: "2025-05-15T10:30:00Z",
-        },
-        {
-          id: 2,
-          metabot_id: 1,
-          prompt:
-            "Show me the customer acquisition trends over the last 6 months",
-          model: "model" as const,
-          model_id: 2,
-          model_name: "Customer Acquisition Trend Analyzer",
-          created_at: "2025-05-15T11:15:00Z",
-          updated_at: "2025-05-15T11:15:00Z",
-        },
-        {
-          id: 3,
-          metabot_id: 1,
-          prompt: "What are our top performing products by sales volume?",
-          model: "metric" as const,
-          model_id: 3,
-          model_name: "Product Performance Ranking",
-          created_at: "2025-05-15T14:22:00Z",
-          updated_at: "2025-05-16T09:45:00Z",
-        },
-      ];
-      setup({ promptSuggestions: prompts });
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-      );
-
-      // should render prompts
-      expect(
-        await screen.findByTestId("metabot-prompt-suggestions"),
-      ).toBeInTheDocument();
-      expect(await screen.findByText(prompts[0].prompt)).toBeInTheDocument();
-      const prompt1 = await screen.findByText(prompts[1].prompt);
-      expect(prompt1).toBeInTheDocument();
-
-      // user should be able to click prompts to start a new convo
-      await userEvent.click(prompt1);
-      await waitFor(async () => {
-        expect(
-          fetchMock.calls(`path:/api/ee/metabot-v3/v2/agent`),
-        ).toHaveLength(1);
+    it("should render markdown for metabot's replies", async () => {
+      setup();
+      mockAgentEndpoint({
+        textChunks: [
+          `0:"# You, but don't tell anyone."`,
+          `2:{"type":"state","version":1,"value":{"queries":{}}}`,
+          `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`,
+        ],
       });
+      await enterChatMessage("Who is your favorite?");
 
-      // unclicked prompts should be gone, but clicked prompt should be in convo
-      expect(await screen.findByText(prompts[1].prompt)).toBeInTheDocument();
-      expect(screen.queryByText(prompts[0].prompt)).not.toBeInTheDocument();
-      expect(
-        screen.queryByTestId("metabot-prompt-suggestions"),
-      ).not.toBeInTheDocument();
+      const heading = await screen.findByRole("heading", { level: 1 });
+      expect(heading).toBeInTheDocument();
+      expect(heading).toHaveTextContent(`You, but don't tell anyone.`);
     });
 
-    it("should make a request for new suggested prompts when the conversation is reset", async () => {
-      setup({ promptSuggestions: [] });
-      await waitFor(async () => {
-        expect(
-          fetchMock.calls(
-            `path:/api/ee/metabot-v3/metabot/1/prompt-suggestions`,
-          ),
-        ).toHaveLength(1);
-      });
+    it("should present the user an option to retry a response", async () => {
+      setup();
+      mockAgentEndpoint({ textChunks: whoIsYourFavoriteResponse });
 
-      await userEvent.click(await resetChatButton());
+      await enterChatMessage("Who is your favorite?");
+      const lastMessage = await lastChatMessage();
+      expect(lastMessage).toHaveTextContent(/You, but don't tell anyone./);
+      expect(
+        await within(lastMessage!).findByTestId("metabot-chat-message-retry"),
+      ).toBeInTheDocument();
+    });
 
-      await waitFor(async () => {
-        expect(
-          fetchMock.calls(
-            `path:/api/ee/metabot-v3/metabot/1/prompt-suggestions`,
-          ),
-        ).toHaveLength(2);
+    it("should successfully rewind a response", async () => {
+      setup();
+      mockAgentEndpoint({
+        // add two messages to ensure we rollback both
+        textChunks: [`0:"Let me think..."`, ...whoIsYourFavoriteResponse],
       });
+      await enterChatMessage("Who is your favorite?");
+
+      const beforeMessages = await screen.findByTestId("metabot-chat-messages");
+      expect(beforeMessages).toHaveTextContent(/Let me think.../);
+      expect(beforeMessages).toHaveTextContent(/You, but don't tell anyone./);
+
+      mockAgentEndpoint({
+        textChunks: [
+          `0:"The answer is always you."`,
+          `2:{"type":"state","version":1,"value":{"queries":{}}}`,
+          `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`,
+        ],
+      });
+      await userEvent.click(
+        await screen.findByTestId("metabot-chat-message-retry"),
+      );
+
+      const afterMessages = await screen.findByTestId("metabot-chat-messages");
+
+      // rolled back both messages
+      expect(afterMessages).not.toHaveTextContent(/Let me think.../);
+      expect(afterMessages).not.toHaveTextContent(
+        /You, but don't tell anyone./,
+      );
+
+      // shows the new one
+      expect(afterMessages).toHaveTextContent(/The answer is always you./);
+    });
+
+    it("should not show retry option for error messages", async () => {
+      setup();
+      fetchMock.post(`path:/api/ee/metabot-v3/v2/agent-streaming`, 500);
+
+      await enterChatMessage("Who is your favorite?");
+
+      const lastMessage = await lastChatMessage();
+      expect(lastMessage).toHaveTextContent(METABOT_ERR_MSG.agentOffline);
+      expect(
+        within(lastMessage!).queryByTestId("metabot-chat-message-retry"),
+      ).not.toBeInTheDocument();
     });
 
     it("should be able to set the prompt input's value from anywhere in the app", async () => {
@@ -360,21 +356,88 @@ describe("metabot", () => {
       expect(await input()).toHaveValue("TEST VAL");
     });
 
-    it("should not present the user an option to retry a prompt (not yet supported)", async () => {
-      setup();
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-      );
+    describe("prompt-suggestions", () => {
+      it("should provide prompt suggestions if available", async () => {
+        const prompts = [
+          {
+            id: 1,
+            metabot_id: 1,
+            prompt: "What is the total revenue for this quarter?",
+            model: "metric" as const,
+            model_id: 1,
+            model_name: "Quarterly Revenue Calculator",
+            created_at: "2025-05-15T10:30:00Z",
+            updated_at: "2025-05-15T10:30:00Z",
+          },
+          {
+            id: 2,
+            metabot_id: 1,
+            prompt:
+              "Show me the customer acquisition trends over the last 6 months",
+            model: "model" as const,
+            model_id: 2,
+            model_name: "Customer Acquisition Trend Analyzer",
+            created_at: "2025-05-15T11:15:00Z",
+            updated_at: "2025-05-15T11:15:00Z",
+          },
+          {
+            id: 3,
+            metabot_id: 1,
+            prompt: "What are our top performing products by sales volume?",
+            model: "metric" as const,
+            model_id: 3,
+            model_name: "Product Performance Ranking",
+            created_at: "2025-05-15T14:22:00Z",
+            updated_at: "2025-05-16T09:45:00Z",
+          },
+        ];
+        setup({ promptSuggestions: prompts });
+        const agentSpy = mockAgentEndpoint({
+          textChunks: whoIsYourFavoriteResponse,
+        });
 
-      await enterChatMessage("Who is your favorite?");
-      const lastMessage = await lastChatMessage();
-      expect(lastMessage).toHaveTextContent(
-        /You are... but don't tell anyone!/,
-      );
-      expect(
-        within(lastMessage!).queryByTestId("metabot-chat-message-retry"),
-      ).not.toBeInTheDocument();
+        // should render prompts
+        expect(
+          await screen.findByTestId("metabot-prompt-suggestions"),
+        ).toBeInTheDocument();
+        expect(await screen.findByText(prompts[0].prompt)).toBeInTheDocument();
+        const prompt1 = await screen.findByText(prompts[1].prompt);
+        expect(prompt1).toBeInTheDocument();
+
+        // user should be able to click prompts to start a new convo
+        await userEvent.click(prompt1);
+        await waitFor(async () => {
+          expect(agentSpy).toHaveBeenCalledTimes(1);
+        });
+
+        // unclicked prompts should be gone, but clicked prompt should be in convo
+        expect(await screen.findByText(prompts[1].prompt)).toBeInTheDocument();
+        expect(screen.queryByText(prompts[0].prompt)).not.toBeInTheDocument();
+        expect(
+          screen.queryByTestId("metabot-prompt-suggestions"),
+        ).not.toBeInTheDocument();
+      });
+
+      it("should make a request for new suggested prompts when the conversation is reset", async () => {
+        setup({ promptSuggestions: [] });
+        await waitFor(async () => {
+          expect(
+            fetchMock.calls(
+              `path:/api/ee/metabot-v3/metabot/1/prompt-suggestions`,
+            ),
+          ).toHaveLength(1);
+        });
+
+        await userEvent.click(await resetChatButton());
+
+        await waitFor(async () => {
+          expect(
+            fetchMock.calls(
+              `path:/api/ee/metabot-v3/metabot/1/prompt-suggestions`,
+            ),
+          ).toHaveLength(2);
+        });
+      });
     });
   });
 
@@ -387,10 +450,9 @@ describe("metabot", () => {
 
     it("should properly send chat messages", async () => {
       setup();
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-        { delay: 50 }, // small delay to cause loading state
+
+      mockAgentEndpoint(
+        { textChunks: whoIsYourFavoriteResponse, initialDelay: 50 }, // small delay to cause loading state
       );
 
       await enterChatMessage("Who is your favorite?", false);
@@ -399,85 +461,277 @@ describe("metabot", () => {
       await enterChatMessage("Who is your favorite?");
       expect(await responseLoader()).toBeInTheDocument();
       expect(
-        await screen.findByText("You are... but don't tell anyone!"),
+        await screen.findByText("You, but don't tell anyone."),
       ).toBeInTheDocument();
 
       // should auto-clear input + refocus
       expect(await input()).toHaveValue("");
       expect(await input()).toHaveFocus();
     });
+
+    it("should properly handle partial messages", async () => {
+      setup();
+
+      const [pause1] = createPauses(2);
+      mockAgentEndpoint({
+        stream: createMockReadableStream(
+          (async function* () {
+            yield `0:"You, but "\n`;
+            await pause1.promise;
+            yield `0:"don't tell anyone."\n`;
+            yield `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`;
+          })(),
+        ),
+      });
+
+      await enterChatMessage("Who is your favorite?");
+      await assertConversation([
+        ["user", "Who is your favorite?"],
+        ["agent", "You, but"],
+      ]);
+
+      pause1.resolve();
+
+      await assertConversation([
+        ["user", "Who is your favorite?"],
+        ["agent", "You, but don't tell anyone."],
+      ]);
+    });
+  });
+
+  describe("tool calls", () => {
+    it("should show list each tool being called as it comes in and cleared when finished", async () => {
+      setup();
+
+      const [pause1, pause2] = createPauses(2);
+      mockAgentEndpoint({
+        stream: createMockReadableStream(
+          (async function* () {
+            yield `9:{"toolCallId":"x","toolName":"analyze_data","args":""}\n`;
+            yield `a:{"toolCallId":"x","result":""}\n`;
+            await pause1.promise;
+            yield `9:{"toolCallId":"y","toolName":"analyze_chart","args":""}\n`;
+            yield `a:{"toolCallId":"y","result":""}\n`;
+            await pause2.promise;
+            yield `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`;
+          })(),
+        ),
+      });
+
+      await enterChatMessage("Analyze this query");
+
+      // should show first tool call
+      expect(await screen.findByText("Analyzing the data")).toBeInTheDocument();
+      expect(
+        screen.queryByText("Inspecting the visualization"),
+      ).not.toBeInTheDocument();
+
+      pause1.resolve();
+
+      // should show both tool calls
+      expect(await screen.findByText("Analyzing the data")).toBeInTheDocument();
+      expect(
+        await screen.findByText("Inspecting the visualization"),
+      ).toBeInTheDocument();
+
+      pause2.resolve();
+
+      // should clear tool call notifications at end of request
+      await waitFor(() => {
+        expect(
+          screen.queryByText("Analyzing the data"),
+        ).not.toBeInTheDocument();
+      });
+      await waitFor(() => {
+        expect(
+          screen.queryByText("Inspecting the visualization"),
+        ).not.toBeInTheDocument();
+      });
+    });
+
+    it("should clear out list of tool calls when new text comes in", async () => {
+      setup();
+
+      const [pause1, pause2, pause3] = createPauses(3);
+      mockAgentEndpoint({
+        stream: createMockReadableStream(
+          (async function* () {
+            yield `9:{"toolCallId":"x","toolName":"analyze_data","args":""}\n`;
+            yield `a:{"toolCallId":"x","result":""}\n`;
+            await pause1.promise;
+            yield `0:"Hey."`;
+            await pause2.promise;
+            yield `9:{"toolCallId":"y","toolName":"analyze_chart","args":""}\n`;
+            yield `a:{"toolCallId":"y","result":""}\n`;
+            await pause3.promise;
+            yield `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`;
+          })(),
+        ),
+      });
+
+      await enterChatMessage("Analyze this query");
+
+      // should show first tool call
+      expect(await screen.findByText("Analyzing the data")).toBeInTheDocument();
+      expect(
+        screen.queryByText("Inspecting the visualization"),
+      ).not.toBeInTheDocument();
+
+      pause1.resolve();
+
+      // should not show any tool calls, only the new text
+      await waitFor(() => {
+        expect(
+          screen.queryByText("Analyzing the data"),
+        ).not.toBeInTheDocument();
+      });
+      expect(screen.getByText("Hey.")).toBeInTheDocument();
+      await waitFor(() => {
+        expect(
+          screen.queryByText("Inspecting the visualization"),
+        ).not.toBeInTheDocument();
+      });
+
+      pause2.resolve();
+
+      // expect text and new tool call to be in the document
+      expect(await screen.findByText("Hey.")).toBeInTheDocument();
+      expect(
+        await screen.findByText("Inspecting the visualization"),
+      ).toBeInTheDocument();
+      expect(screen.queryByText("Analyzing the data")).not.toBeInTheDocument();
+
+      pause3.resolve();
+
+      await waitFor(async () => {
+        await assertConversation([
+          ["user", "Analyze this query"],
+          ["agent", "Hey."],
+        ]);
+      });
+    });
+
+    it("should start a new message if there's tool calls between streamed text parts", async () => {
+      setup();
+      mockAgentEndpoint(
+        {
+          textChunks: [
+            `0:"Response 1"`,
+            `9:{"toolCallId":"x","toolName":"x","args":""}`,
+            `a:{"toolCallId":"x","result":""}`,
+            `0:"Response 2"`,
+            `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`,
+          ],
+        }, // small delay to cause loading state
+      );
+      await enterChatMessage("Request");
+
+      await assertConversation([
+        ["user", "Request"],
+        ["agent", "Response 1"],
+        ["agent", "Response 2"],
+      ]);
+    });
   });
 
   describe("errors", () => {
     it("should handle service error response", async () => {
       setup();
-      fetchMock.post(`path:/api/ee/metabot-v3/v2/agent`, 500);
+      fetchMock.post(`path:/api/ee/metabot-v3/v2/agent-streaming`, 500);
 
       await enterChatMessage("Who is your favorite?");
 
-      expect(await lastChatMessage()).toHaveTextContent(
-        METABOT_ERR_MSG.agentOffline,
-      );
+      await assertConversation([
+        ["user", "Who is your favorite?"],
+        ["agent", METABOT_ERR_MSG.agentOffline],
+      ]);
       expect(await input()).toHaveValue("Who is your favorite?");
     });
 
     it("should handle non-successful responses", async () => {
       setup();
-      fetchMock.post(`path:/api/ee/metabot-v3/v2/agent`, 400);
+      fetchMock.post(`path:/api/ee/metabot-v3/v2/agent-streaming`, 400);
 
       await enterChatMessage("Who is your favorite?");
 
-      expect(await lastChatMessage()).toHaveTextContent(
-        METABOT_ERR_MSG.default,
-      );
+      await assertConversation([
+        ["user", "Who is your favorite?"],
+        ["agent", METABOT_ERR_MSG.default],
+      ]);
+      expect(await input()).toHaveValue("Who is your favorite?");
+    });
+
+    it("should handle show error if data error part is in response", async () => {
+      setup();
+      mockAgentEndpoint({ textChunks: erroredResponse });
+
+      await enterChatMessage("Who is your favorite?");
+
+      await assertConversation([
+        ["user", "Who is your favorite?"],
+        ["agent", METABOT_ERR_MSG.default],
+      ]);
       expect(await input()).toHaveValue("Who is your favorite?");
     });
 
     it("should not show a user error when an AbortError is triggered", async () => {
       setup();
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-        { delay: 50 }, // small delay to cause loading state
-      );
+      mockAgentEndpoint({ textChunks: whoIsYourFavoriteResponse });
 
       await enterChatMessage("Who is your favorite?");
 
-      expect(await chatMessages()).toHaveLength(1);
+      await assertConversation([
+        ["user", "Who is your favorite?"],
+        ["agent", "You, but don't tell anyone."],
+      ]);
+
       await userEvent.click(await resetChatButton());
-      await waitFor(() => {
-        expect(
-          screen.queryByTestId("metabot-chat-message"),
-        ).not.toBeInTheDocument();
-      });
+
+      await assertConversation([]);
       expect(await input()).toHaveValue("");
+    });
+
+    it("should remove previous error messages and prompt when submiting next prompt", async () => {
+      setup();
+      fetchMock.post(`path:/api/ee/metabot-v3/v2/agent-streaming`, 500);
+
+      await enterChatMessage("Who is your favorite?");
+
+      await assertConversation([
+        ["user", "Who is your favorite?"],
+        ["agent", METABOT_ERR_MSG.agentOffline],
+      ]);
+      expect(await input()).toHaveValue("Who is your favorite?");
+
+      mockAgentEndpoint({
+        textChunks: whoIsYourFavoriteResponse,
+      });
+      await enterChatMessage("Who is your favorite?");
+      await assertConversation([
+        ["user", "Who is your favorite?"],
+        ["agent", "You, but don't tell anyone."],
+      ]);
     });
   });
 
   describe("context", () => {
     it("should send along default context", async () => {
       setup();
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-      );
+      const agentSpy = mockAgentEndpoint({
+        textChunks: whoIsYourFavoriteResponse,
+      });
 
       await enterChatMessage("Who is your favorite?");
 
       expect(
         isMatching(
           { current_time_with_timezone: P.string },
-          (await lastReqBody())?.context,
+          (await lastReqBody(agentSpy))?.context,
         ),
       ).toEqual(true);
     });
 
     it("should allow components to register additional context", async () => {
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-      );
-
       const TestComponent = () => {
         useRegisterMetabotContextProvider(
           () =>
@@ -498,6 +752,10 @@ describe("metabot", () => {
         ),
       });
 
+      const agentSpy = mockAgentEndpoint({
+        textChunks: whoIsYourFavoriteResponse,
+      });
+
       await enterChatMessage("Who is your favorite?");
 
       expect(
@@ -506,91 +764,147 @@ describe("metabot", () => {
             current_time_with_timezone: P.string,
             user_is_viewing: [{ type: "dashboard", id: 1 }],
           },
-          (await lastReqBody())?.context,
+          (await lastReqBody(agentSpy))?.context,
         ),
       ).toBe(true);
     });
   });
 
+  describe("convo state", () => {
+    it("should update the convo state on a successful request", async () => {
+      const { store } = setup();
+      // TODO: make enterprise store
+      const getState = () => getMetabotState(store.getState() as any);
+
+      mockAgentEndpoint({
+        stream: createMockReadableStream(
+          (async function* () {
+            yield `2:{"type":"state","version":1,"value":{"queries":{}}}\n`;
+            // assert that state hasn't been updated mid-response
+            expect(getState()).toEqual({});
+            yield `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`;
+          })(),
+        ),
+      });
+
+      expect(getState()).toEqual({});
+      await enterChatMessage("Request");
+      expect(getState()).toEqual({ queries: {} });
+    });
+
+    it("should not update the convo state on a failed request", async () => {
+      const { store } = setup();
+      // TODO: make enterprise store
+      const getState = () => getMetabotState(store.getState() as any);
+
+      mockAgentEndpoint({
+        textChunks: [
+          `2:{"type":"state","version":1,"value":{"queries":{}}}`,
+          ...erroredResponse,
+        ],
+      });
+
+      expect(getState()).toEqual({});
+      await enterChatMessage("Request");
+      expect(getState()).toEqual({});
+    });
+  });
+
   describe("history", () => {
-    it("should send history from last response along with next message", async () => {
+    it("should send conversation history along with future messages", async () => {
       setup();
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-      );
 
-      // send a message to get some history back
+      // create some chat history + wait for a response
+      mockAgentEndpoint({
+        textChunks: whoIsYourFavoriteResponse,
+        initialDelay: 50,
+      });
       await enterChatMessage("Who is your favorite?");
-      await waitFor(() =>
-        expect(
-          fetchMock.calls(`path:/api/ee/metabot-v3/v2/agent`),
-        ).toHaveLength(1),
-      );
+      expect(
+        await screen.findByText("You, but don't tell anyone."),
+      ).toBeInTheDocument();
 
-      // send another message and check that there is proper history
+      // send another message and check that the proper history is being sent along
+      const agentSpy = mockAgentEndpoint({
+        textChunks: [], // next response doesn't matter
+      });
       await enterChatMessage("Hi!");
-      const lastCall = fetchMock.lastCall();
-      const bodyStr = String(await lastCall?.[1]?.body) || "";
-      const sentHistory = JSON.parse(bodyStr)?.history;
-      expect(sentHistory).toEqual(whoIsYourFavoriteResponse.history);
+      const reqBody = await lastReqBody(agentSpy);
+      expect(reqBody?.history).toEqual([
+        { content: "Who is your favorite?", role: "user" },
+        { content: "You, but don't tell anyone.", role: "assistant" },
+      ]);
     });
 
     it("should not clear history when metabot is hidden or opened", async () => {
       const { store } = setup();
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-      );
+      const agentSpy = mockAgentEndpoint({
+        textChunks: whoIsYourFavoriteResponse,
+      });
 
       // send a message to get some history back
       await enterChatMessage("Who is your favorite?");
-      await waitFor(() =>
-        expect(
-          fetchMock.calls(`path:/api/ee/metabot-v3/v2/agent`),
-        ).toHaveLength(1),
-      );
+      await waitFor(() => expect(agentSpy).toHaveBeenCalledTimes(1));
 
       // close, open, and then send another message and check that there is no history
       hideMetabot(store.dispatch);
       showMetabot(store.dispatch);
       await enterChatMessage("Hi!");
-      const lastCall = fetchMock.lastCall();
-      const bodyStr = String(await lastCall?.[1]?.body) || "";
-      const sentHistory = JSON.parse(bodyStr)?.history;
-      expect(sentHistory).toEqual(whoIsYourFavoriteResponse.history);
+      const lastCall = agentSpy.mock.lastCall;
+      const reqBody = JSON.parse(String(lastCall?.[1]?.body));
+      const sentHistory = reqBody.history;
+      expect(sentHistory).not.toEqual([]);
+    });
+
+    it("should merge text chunks in the history", async () => {
+      const { store } = setup();
+      mockAgentEndpoint({
+        textChunks: [
+          `0:"You, but "`,
+          `0:"don't tell anyone."`,
+          `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`,
+        ],
+      });
+
+      const initialHistory = getHistory(store.getState() as any);
+      expect(initialHistory).toEqual([]);
+
+      await enterChatMessage("Who is your favorite?");
+
+      const finalHistory = getHistory(store.getState() as any);
+      expect(finalHistory).toHaveLength(2);
+      expect(finalHistory[0].role).toBe("user");
+      expect(finalHistory[0].content).toBe("Who is your favorite?");
+      expect(finalHistory[1].role).toBe("assistant");
+      expect(finalHistory[1].content).toBe("You, but don't tell anyone.");
     });
 
     it("should clear history when the user hits the reset button", async () => {
       const { store } = setup();
-      fetchMock.post(
-        `path:/api/ee/metabot-v3/v2/agent`,
-        whoIsYourFavoriteResponse,
-      );
+      const getState = () => getMetabot(store.getState() as any);
+      mockAgentEndpoint({ textChunks: whoIsYourFavoriteResponse });
 
       // send a message to get some history back
       await enterChatMessage("Who is your favorite?");
-      await waitFor(() =>
-        expect(
-          fetchMock.calls(`path:/api/ee/metabot-v3/v2/agent`),
-        ).toHaveLength(1),
-      );
+      await assertConversation([
+        ["user", "Who is your favorite?"],
+        ["agent", "You, but don't tell anyone."],
+      ]);
 
-      const beforeResetState = getMetabotState(store);
+      const beforeResetState = getState();
       expect(beforeResetState.conversationId).not.toBe(null);
-      expect(beforeResetState.messages).toHaveLength(2);
       expect(_.omit(beforeResetState.messages[0], "id")).toStrictEqual({
         role: "user",
         message: "Who is your favorite?",
       });
       expect(_.omit(beforeResetState.messages[1], "id")).toStrictEqual({
         role: "agent",
-        message: "You are... but don't tell anyone!",
+        message: "You, but don't tell anyone.",
       });
 
       await userEvent.click(await resetChatButton());
 
-      const afterResetState = getMetabotState(store);
+      const afterResetState = getState();
       expect(afterResetState.conversationId).not.toBe(
         beforeResetState.conversationId,
       );
@@ -630,38 +944,13 @@ describe("metabot", () => {
   });
 });
 
-const whoIsYourFavoriteResponse = {
-  reactions: [
-    {
-      type: "metabot.reaction/message",
-      message: "You are... but don't tell anyone!",
-    },
-  ],
-  history: [
-    {
-      role: "user",
-      content: "Who is your favorite?",
-    },
-    {
-      content: "",
-      role: "assistant",
-      "tool-calls": [
-        {
-          id: "call_PVmnR8mcnYFF2AmqupSKzJDh",
-          name: "who-is-your-favorite",
-          arguments: {},
-        },
-      ],
-    },
-    {
-      role: "tool",
-      "tool-call-id": "call_PVmnR8mcnYFF2AmqupSKzJDh",
-      content: "You are... but don't tell anyone!",
-    },
-    {
-      content: "You are... but don't tell anyone!",
-      role: "assistant",
-    },
-  ],
-  state: {},
-};
+const whoIsYourFavoriteResponse = [
+  `0:"You, but don't tell anyone."`,
+  `2:{"type":"state","version":1,"value":{"queries":{}}}`,
+  `d:{"finishReason":"stop","usage":{"promptTokens":4916,"completionTokens":8}}`,
+];
+
+const erroredResponse = [
+  `3:{"type": "error", "value": "Couldn't do the thing"}`,
+  `d:{"finishReason":"error","usage":{"promptTokens":4916,"completionTokens":8}}`,
+];
